+++ lib/rivet-api/src/models/cloud_upload_prepare_file.rs
@@ -16,7 +16,7 @@
 pub struct CloudUploadPrepareFile {
     /// Unsigned 64 bit integer.
     #[serde(rename = "content_length", skip_serializing_if = "Option::is_none")]
-    pub content_length: Option<f64>,
+    pub content_length: Option<i64>,
     /// The MIME type of the file.
     #[serde(rename = "content_type", skip_serializing_if = "Option::is_none")]
     pub content_type: Option<String>,
